Move data-studio FE module to the feature tier
